==> R/00_all.qmd <==
---
title: "alltest"
format: html
editor: visual
---

Library

```{r}
library("knitr")
```

```{r}
#| child:
#| - "01_load.qmd"
#| - "02_clean.qmd"
#| - "03_augment.qmd"
#| - "04_describe.qmd"
#| - "05_analysis_1.qmd"
#| - "06_analysis_2.qmd"
#| - "07_analysis_3.qmd"
#| - "08_analysis_5.qmd"
#| - "09_analysis_6.qmd"
#| - "10_analysis_7.qmd"
```
