==> R/02_clean.qmd <==
---
title: "data_cleaning"
format: html
edtiro: visual
---

# Libraries & Functions

```{r}
#| warning: false
library("tidyverse")
library("dplyr")
library("janitor")

source("../R/99_proj_func.R")
```

# General cleaning guidelines for this project

-   Subsetting columns that we're not going to use for this project.

-   Cleaning column names with the `janitor` library.

-   Skip over the first and/or second lines of the excel files. (This is done on `data_loading.qmd` to have the appropiate data structure from the start).

-   Drop NAs (Check with the NA_check function if it's necessary or harmful for the data)

# Cleaning table 2

```{r}
#| echo: false
dirty_2 <-read_tsv(file = "../data/dirty_data/supp_2.tsv", show_col_types = FALSE)

#Keeping the columns needed for the analysis

clean_2 <- dirty_2 |>
  select(
    gene          = 1,
    protein       = 2,
    logLFQ_astro  = 3,
    logLFQ_micro  = 4,
    logLFQ_neuron = 5,
    logLFQ_oligo  = 6
  ) |>
  slice(-(1:2)) 


clean_2 <- clean_2 |>
  mutate(
    logLFQ_astro  = as.numeric(logLFQ_astro),
    logLFQ_micro  = as.numeric(logLFQ_micro),
    logLFQ_neuron = as.numeric(logLFQ_neuron),
    logLFQ_oligo  = as.numeric(logLFQ_oligo)
  )

```

# Cleaning table 3

```{r}
#| echo: false

dirty_3 <- read_tsv(file = "../data/dirty_data/supp_3.tsv", show_col_types = FALSE)

#Janitor
clean_3 <- dirty_3 |>
  clean_names()

#Moving gene name to the front
clean_3 <- clean_3 |>
  relocate(gene_name)

#Checking NAs
print(
  paste("The total % of NA values are: ", round(NA_check(clean_3), 2))
)
```

```{r}
#Transforming NaNs into NAs to drop them and cleaning column names
clean_3 <- clean_3 |>
  mutate(across(everything(),
                ~ replace(., is.nan(.), NA))) |>
           drop_na()

#Important note on removing NAs:
#NAs compose 22.64% of the total values, however drop_na() removes ANY rows containing NAs
#Therefore a substantial data loss is obtained and data analysis should be conducted with caution

rows_in_dirty <- nrow(dirty_3) #19501 entries
rows_in_clean <- nrow(clean_3) #11062 entries


pct_lost <- (rows_in_dirty - rows_in_clean) / rows_in_dirty *100

print(
  paste("The total % of lost rows are", round(pct_lost, 2))
)
```

# Cleaning table 5

```{r}
#| echo: false

dirty_5 <- read_tsv(file = "../data/dirty_data/supp_5.tsv", show_col_types = FALSE)

#Dropping duplicate gene_names (keeping the first one)

clean_5 <- dirty_5 |> 
  select(-contains("young"), -PEP, - `Mol. weight [kDa]`, -`Protein IDs`, -`Majority protein IDs`, -`Sequence coverage [%]`) |> 
  clean_names() |> 
  distinct(gene_names, .keep_all = T) 

NA_check(clean_5) #0.32% of the values are NAs. So dropping NAs won't affect our results. 

clean_5 <- clean_5 |> 
  drop_na()

```

# Cleaning table 17

```{r}
dirty_17 <- read_tsv(file = "../data/dirty_data/supp_17.tsv", show_col_types = FALSE)

clean_17 <- dirty_17 |> 
  clean_names() 

NA_check(clean_17) #17.55% of the data is NAs
#BUT if an NA check is conducted on the subset that will be used afterwards for the analysis, it doesn't have any NAs!

NA_check(clean_17 |> 
           select(14:23)) #0% NAs!!

#So for this specific table it was decided that the NA drop can be skipped, since it would only hurt the analysis. 

```

# Writing clean data to tables

```{r}
#Create directory in case it doesn't exist already 
dir.create("../data/clean_data", recursive = TRUE, showWarnings = FALSE)

write_tsv(clean_2, "../data/clean_data/clean_2.tsv")

write_tsv(clean_3, "../data/clean_data/clean_3.tsv")

write_tsv(clean_5, "../data/clean_data/clean_5.tsv")

write_tsv(clean_17, "../data/clean_data/clean_17.tsv")
```
