==> R/08_analysis_5.qmd <==
---
title: "Log2"
format: html
---

```{r}
library(tidyverse)
library(ggrepel)
library(patchwork)

#Sourcing function to draw panels
source("99_proj_func.R")

all_df <- read_tsv("../data/augmented_data/table_2_log.tsv", show_col_types = FALSE)

#Selecting only the rows corresponding to marker genes.
#These points will be labeled and colored in the plot using geom_text_repel()

labels_df <- all_df |>
  group_by(type) |>
  filter(
    (type == "Astrocytes"       & gene %in% astro_markers)  |
    (type == "Microglia"        & gene %in% micro_markers)  |
    (type == "Neurons"          & gene %in% neuron_markers) |
    (type == "Oligodendrocytes" & gene %in% oligo_markers)
  ) |>
  ungroup()

#Customing color palette for each marker category

cols <- c(
  "Astrocyte markers"       = "#1A9850",
  "Microglia markers"       = "#2B83BA",
  "Neuron markers"          = "#00008B",
  "Oligodendrocyte markers" = "#D7191C",
  "Other"                   = "grey70"
)

# Base theme applied to all panels
base_theme <- theme_bw(base_size = 12) +
  theme(
    plot.title = element_text(face = "bold", size = 14),
    legend.title = element_blank(),
    legend.text  = element_text(size = 14)
  )



# Generating a panel for each cell type
# Each panel uses its subset of data and its marker labels

p_astro  <- draw_panel(all_df |> filter(type == "Astrocytes"),       labels_df |> filter(type == "Astrocytes"),       "Astrocytes")
p_micro  <- draw_panel(all_df |> filter(type == "Microglia"),        labels_df |> filter(type == "Microglia"),        "Microglia")
p_neuron <- draw_panel(all_df |> filter(type == "Neurons"),          labels_df |> filter(type == "Neurons"),          "Neurons")
p_oligo  <- draw_panel(all_df |> filter(type == "Oligodendrocytes"), labels_df |> filter(type == "Oligodendrocytes"), "Oligodendrocytes")

# Combining all four panels into a single figure
# Arranging them in 2 rows and place the legend at the bottom

combined <- p_astro + p_micro + p_neuron + p_oligo +
  plot_layout(nrow = 2, guides = "collect") &
  theme(legend.position = "bottom")

ggsave("../results/LogFoldChange_cell_types.jpg", 
       combined, 
       width = 18,
       height = 10, 
       dpi = 300)

```
